django/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Pillow==9.2.0
pip==21.2.3
PyMySQL==1.0.2
pytz==2022.2.1
setuptools==57.4.0
sqlparse==0.4.2
